backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4
dj-database-url>=2.1.0

# LangChain & AI
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
langchain-chroma>=0.1.0
google-generativeai>=0.3.2

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.0
python-magic>=0.4.27

# Embeddings (Open Source)
sentence-transformers>=2.2.2
chromadb>=0.4.22

# Async & Background Tasks
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Real-time
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Utilities
python-dotenv>=1.0.0
python-decouple>=3.8
whitenoise>=6.6.0
gunicorn>=21.2.0
cloudinary>=1.38.0
django-cloudinary-storage>=0.3.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.2
factory-boy>=3.3.0
faker>=22.0.0

# Development
black>=23.12.0
flake8>=6.1.0
isort>=5.13.2

# Monitoring & Logging
sentry-sdk>=1.38.0
django-extensions>=3.2.3
